+++ Escape_V2/Escape_V2/GuardPatrolState.cpp
@@ -36,7 +36,7 @@ void GuardPatrolState::Init(int number, sf::Vector2f* p_position, AnimatedSprite
 
 	mp_position = &Settings::ms_guards.at(m_number);
 
-	for(int i = 0; i < Settings::m_allGuardWaypoints[m_number].size(); i++) {
+	for(unsigned int i = 0; i < Settings::m_allGuardWaypoints[m_number].size(); i++) {
 	
 		m_waypoints.push_back(Settings::m_allGuardWaypoints[m_number].at(i));
 
@@ -73,7 +73,7 @@ void GuardPatrolState::Cleanup() {
 
 void GuardPatrolState::UpdateAnimation() {
 	mp_sprite->Update();
-	m_rotation = atan2(mp_position->y - m_waypoints.at(m_nextWaypoint).y, mp_position->x - m_waypoints.at(m_nextWaypoint).x) * 180 / M_PI - 90;
+	m_rotation = atan2(mp_position->y - m_waypoints.at(m_nextWaypoint).y, mp_position->x - m_waypoints.at(m_nextWaypoint).x) * 180 / static_cast<float>(M_PI) - 90;
 	mp_sprite->getSprite()->setRotation(m_rotation);
 }
